requirements.txt
dnspython
pandas
openpyxl
beautifulsoup4
requests
streamlit
nest-asyncio
xlsxwriter
python-dotenv
python-telegram-bot